test(`cqn4sql`): refactor calculated elements test suite
Split the monolithic calculated-elements.test.js (1,254 lines, 76 tests)
into a calculated-elements.test/ directory with topic-based spec files,
mirroring the pattern from assoc2join.test/, exists.test/, and
nested-projections.test/.

- select-list.spec.js  (50 tests)
- other-places.spec.js (17 tests)
- localized.spec.js    (4 tests)
- exists.spec.js       (2 tests)
- drafts.spec.js       (2 tests)
- stored.spec.js       (1 test)

Adopt the shared loadModel/expectCqn helpers, rename query → transformed,
and indent cds.ql template literals to match the surrounding JS context.
The localized describe verifies the SELECT.localized flag with a separate
standard assertion since expectCqn's normalize drops inherited enumerable
props.
